wheel24Deals/requirements.txt
asgiref==3.8.1
Django==4.2.19
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
psycopg2==2.9.10
sqlparse==0.5.3
typing-extensions==4.12.2
tzdata==2025.1
whitenoise==6.9.0
